refactor(x): rename `sendMail` module
